Relocate python test + UI quality fallback (Phase 3, batch 11)
Issue #683 (parent #680, Phase 3) batch 11: migrate the
`maybe_materialize_python_test_fallback` and
`maybe_apply_deterministic_quality_fallback` materializers from
`turn.rs` to `scaffold_pipeline.rs` via the veneer pattern.

Migrated as free fns in `scaffold_pipeline.rs` (`pub(super)`):

* `maybe_materialize_python_test_fallback(&mut Agent)
    -> Result<Option<String>, String>`
* `maybe_apply_deterministic_quality_fallback(
    &Agent, request, relative_target) -> Result<bool, String>`

Promoted in `turn.rs` to `pub(super)`:

* `Agent::maybe_apply_deterministic_framework_support_files`
* `Agent::maybe_apply_requested_port_script`

(Both stay as Agent methods this PR; promotion is needed so the
moved `maybe_apply_deterministic_quality_fallback` free fn can call
them via `agent.<method>()` across the module boundary.)

`turn.rs` adjustments:

* 2 migrated method slots stay as one-line veneer shims so caller
  sites (`run_actor_loop` repair paths) stay unchanged.
* Pruned now-unused `EVENT_DETERMINISTIC_PYTHON_TEST_FALLBACK` from
  the top-of-file `super::scaffold_pipeline::{...}` import block.

LOC delta:

* `turn.rs`: 30,765 → 30,648 (-117 LOC)
* `scaffold_pipeline.rs`: 1,255 → 1,396 (+141 LOC)
* Cumulative `turn.rs`: 39,489 → 30,648 (-8,841 LOC, -22.4%)

`pub(super)` limited / no facade re-export (DR3-001).
`turn.rs` remains the only in-crate consumer.

Verification:

* `cargo fmt`
* `cargo clippy --lib --all-targets -- -D warnings` clean
* `cargo test --lib` 3,114 passed / 0 failed
